requirements.txt
asgiref==3.8.1
captcha==0.6.0
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
utils==1.0.2
views==0.3
